+++ src/pages/AllBuyNowPage.js
@@ -1,9 +1,9 @@
-import AllPageTemplate from "../../components/AllPageTemplate";
-import { db } from "../../config/firebase"
+import AllPageTemplate from "../components/AllPageTemplate";
+import { db } from "../config/firebase"
 import {collection, getDocs} from "firebase/firestore";
 import { useState, useEffect } from "react"
 
-const AllBidPage = () => {
+const AllBuyNowPage = () => {
     const listingRef = collection(db, "listing")
     const [listing, setListing] = useState([])
     const fetchedData = () => {
@@ -17,11 +17,13 @@ const AllBidPage = () => {
     if (listing !== []) {
         let temp = []
         listing.map((card) => {
-            if (card.product_pricing === "bid now") temp.push(card)
+            if (card.product_pricing === "buy now") temp.push(card)
         })
         return (
-            <AllPageTemplate fetchedData={temp} title={"Bid now"}/>
+            <AllPageTemplate fetchedData={temp} title={"Buy Now"}/>
         )
     }
 }
-export default AllBidPage
+
+
+export default AllBuyNowPage
