+++ src/pages/AllTradingCardsPage.js
@@ -1,9 +1,9 @@
-import AllPageTemplate from "../../components/AllPageTemplate";
-import { db } from "../../config/firebase"
+import AllPageTemplate from "../components/AllPageTemplate";
+import { db } from "../config/firebase"
 import {collection, getDocs} from "firebase/firestore";
 import { useState, useEffect } from "react"
 
-const AllBuyNowPage = () => {
+const AllTradingCardsPage = () => {
     const listingRef = collection(db, "listing")
     const [listing, setListing] = useState([])
     const fetchedData = () => {
@@ -17,13 +17,13 @@ const AllBuyNowPage = () => {
     if (listing !== []) {
         let temp = []
         listing.map((card) => {
-            if (card.product_pricing === "buy now") temp.push(card)
+            if (card.category === "trading card") temp.push(card)
         })
         return (
-            <AllPageTemplate fetchedData={temp} title={"Buy Now"}/>
+            <AllPageTemplate fetchedData={temp} title={"Trading Cards"}/>
         )
     }
 }
 
 
-export default AllBuyNowPage
+export default AllTradingCardsPage
